Move router/index.js to init/router.js
